fix: 별점 평점은 소숫점 한 자리까지 저장하도록 수정

+++ src/main/java/com/hyundai/app/store/mapper/StoreMapper.java
@@ -16,6 +16,6 @@ public interface StoreMapper {
     Store getStoreDetail(@Param("storeId") int storeId);
     List<Hashtag> getPopularHashtagsOfStore(@Param("storeId") int storeId);
     void saveReview(@Param("review") Review review);
-    void updateAvgScore(@Param("storeId") int storeId, @Param("avgScore") float avgScore);
+    void updateAvgScore(@Param("storeId") int storeId, @Param("avgScore") double avgScore);
     void updateReviewCount(@Param("storeId") int storeId);
 }
